[Add] travis auto rebase
