+++ .travis.yml
@@ -3,7 +3,7 @@ node_js:
   - "12"
 branches:
   only:
-  - release
+  - develop
 env:
   - NODE_ENV=test PORT=3100
 install:
@@ -12,3 +12,5 @@ script:
   - npm start &
   - sleep 3
   - npm test
+after_success:
+  - ./travis-merge.sh
