+++ utils/spotify.js
@@ -1,0 +1,18 @@
+const SpotifyWebApi = require("spotify-web-api-node");
+
+// credentials are optional
+const spotifyApi = new SpotifyWebApi({
+    clientId: process.env.CLIENT_ID,
+    clientSecret: process.env.CLIENT_SECRET,
+});
+spotifyApi.clientCredentialsGrant().then(
+    function (data) {
+        // Save the access token so that it's used in future calls
+        spotifyApi.setAccessToken(data.body["access_token"]);
+    },
+    function (err) {
+        console.log("Something went wrong when retrieving an access token", err);
+    }
+);
+
+module.exports = spotifyApi;
